Move configuration cache classes related to _Java Object Serialization_ to separate package
