+++ subprojects/instant-execution/src/main/kotlin/org/gradle/instantexecution/serialization/codecs/Codecs.kt
@@ -42,6 +42,8 @@ import org.gradle.api.tasks.util.internal.PatternSpecFactory
 import org.gradle.execution.plan.TaskNodeFactory
 import org.gradle.initialization.BuildRequestMetaData
 import org.gradle.instantexecution.problems.DocumentationSection.NotYetImplementedJavaSerialization
+import org.gradle.instantexecution.serialization.codecs.jos.SerializableWriteObjectCodec
+import org.gradle.instantexecution.serialization.codecs.jos.SerializableWriteReplaceCodec
 import org.gradle.instantexecution.serialization.codecs.transform.ChainedTransformationNodeCodec
 import org.gradle.instantexecution.serialization.codecs.transform.DefaultTransformerCodec
 import org.gradle.instantexecution.serialization.codecs.transform.InitialTransformationNodeCodec
